Move the custom gql types into separate file
