+++ packages/insomnia-app/app/ui/components/editors/body/graph-ql-editor.tsx
@@ -3,7 +3,7 @@ import classnames from 'classnames';
 import { EditorFromTextArea, TextMarker } from 'codemirror';
 import electron, { OpenDialogOptions } from 'electron';
 import { readFileSync } from 'fs';
-import { GraphQLArgument, GraphQLSchema, GraphQLType } from 'graphql';
+import { GraphQLSchema } from 'graphql';
 import { parse, print, typeFromAST } from 'graphql';
 import { buildClientSchema, getIntrospectionQuery } from 'graphql/utilities';
 import { json as jsonPrettify } from 'insomnia-prettify';
@@ -24,7 +24,8 @@ import type { ResponsePatch } from '../../../../network/network';
 import * as network from '../../../../network/network';
 import { Dropdown, DropdownButton, DropdownDivider, DropdownItem } from '../../base/dropdown';
 import CodeEditor from '../../codemirror/code-editor';
-import GraphqlExplorer, { GraphQLFieldWithOptionalArgs } from '../../graph-ql-explorer/graph-ql-explorer';
+import GraphqlExplorer from '../../graph-ql-explorer/graph-ql-explorer';
+import { ActiveReference } from '../../graph-ql-explorer/graph-ql-types';
 import HelpTooltip from '../../help-tooltip';
 import { showModal } from '../../modals';
 import ResponseDebugModal from '../../modals/response-debug-modal';
@@ -40,12 +41,6 @@ interface GraphQLBody {
   query: string;
   variables?: Record<string, any>;
   operationName?: string;
-}
-
-interface ActiveReference {
-  type: GraphQLType | null;
-  argument: GraphQLArgument | null;
-  field: GraphQLFieldWithOptionalArgs | null;
 }
 
 interface Props {
