+++ packages/insomnia-app/app/ui/components/graph-ql-explorer/graph-ql-default-value.tsx
@@ -1,7 +1,7 @@
 import { astFromValue, print } from 'graphql';
 import React, { PureComponent } from 'react';
 
-import { GraphQLFieldWithParentName } from './graph-ql-explorer';
+import { GraphQLFieldWithParentName } from './graph-ql-types';
 
 interface Props {
   field: GraphQLFieldWithParentName;
