+++ packages/insomnia-app/app/ui/components/graph-ql-explorer/graph-ql-explorer-field-link.tsx
@@ -2,7 +2,7 @@ import { autoBindMethodsForReact } from 'class-autobind-decorator';
 import React, { PureComponent } from 'react';
 
 import { AUTOBIND_CFG } from '../../../common/constants';
-import { GraphQLFieldWithParentName } from './graph-ql-explorer';
+import { GraphQLFieldWithParentName } from './graph-ql-types';
 
 interface Props {
   onNavigate: (type: GraphQLFieldWithParentName) => void;
