+++ packages/insomnia-app/app/ui/components/graph-ql-explorer/graph-ql-explorer-fields-list.tsx
@@ -5,10 +5,10 @@ import React, { FC } from 'react';
 import Tooltip from '../../components/tooltip';
 import MarkdownPreview from '../markdown-preview';
 import GraphQLDefaultValue from './graph-ql-default-value';
-import { GraphQLFieldWithParentName } from './graph-ql-explorer';
 import { GraphQLExplorerArgLinks } from './graph-ql-explorer-arg-links';
 import GraphQLExplorerFieldLink from './graph-ql-explorer-field-link';
 import GraphQLExplorerTypeLink from './graph-ql-explorer-type-link';
+import { GraphQLFieldWithParentName } from './graph-ql-types';
 
 interface Props {
   fields: GraphQLFieldWithParentName[];
